Update `BIN_BASE_*_HEIGH` values to align with spec and output 4.65mm height light baseplate
The current output light baseplate has a 4.5mm height. According to spec https://gridfinity.xyz/specification/, the light baseplate should have a 4.65mm height.

The baseplate section heights should be the following:

* Top section: 2.15mm (`0.265` in code)
* Mid section: 1.8mm (`0.18` in code)
* Bottom Section: 0.7mm (`0.07` in code)

I'm unsure why `BIN_BASE_TOP_SECTION_HEIGH` needs to be set to `0.265` instead of `0.215`, but the result outputs a baseplate with a height of 4.65mm.